DL-3003 Updated to SBT 1.3 and Scala 2.12
