+++ build.sbt
@@ -1,0 +1,42 @@
+import uk.gov.hmrc._
+import DefaultBuildSettings.{defaultSettings, scalaSettings}
+import uk.gov.hmrc.SbtAutoBuildPlugin
+import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin
+import uk.gov.hmrc.versioning.SbtGitVersioning
+import play.sbt.routes.RoutesKeys.routesGenerator
+import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin.publishingSettings
+
+val appName: String = "business-customer-frontend"
+lazy val appDependencies : Seq[ModuleID] = AppDependencies()
+
+lazy val plugins : Seq[Plugins] = Seq(play.sbt.PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin, SbtArtifactory)
+lazy val playSettings : Seq[Setting[_]] = Seq.empty
+lazy val scoverageSettings = {
+  import scoverage.ScoverageKeys
+  Seq(
+    ScoverageKeys.coverageExcludedPackages :=
+      "<empty>;Reverse.*;views.html.*;app.Routes.*;prod.*;uk.gov.hmrc.*;testOnlyDoNotUseInAppConf.*;forms.*;config.*;models.*;views.*",
+    ScoverageKeys.coverageMinimum := 99,
+    ScoverageKeys.coverageFailOnMinimum := true,
+    ScoverageKeys.coverageHighlighting := true
+  )
+}
+lazy val microservice: Project = Project(appName, file("."))
+  .enablePlugins(plugins : _*)
+  .settings(
+    playSettings,
+    routesGenerator := InjectedRoutesGenerator,
+    scalaSettings,
+    publishingSettings,
+    defaultSettings(),
+    scoverageSettings,
+    scalaVersion := "2.12.10",
+    libraryDependencies ++= appDependencies,
+    retrieveManaged := true,
+    evictionWarningOptions in update := EvictionWarningOptions.default.withWarnScalaVersionEviction(false),
+    majorVersion := 4
+  )
+  .settings(
+    resolvers += Resolver.bintrayRepo("hmrc", "releases"),
+    resolvers += Resolver.jcenterRepo
+  )
