--- project/MicroService.scala
@@ -1,49 +1,0 @@
-import play.routes.compiler.InjectedRoutesGenerator
-import sbt.Keys._
-import sbt._
-import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin._
-import uk.gov.hmrc.versioning.SbtGitVersioning.autoImport.majorVersion
-
-trait MicroService {
-
-  import uk.gov.hmrc._
-  import DefaultBuildSettings.{defaultSettings, scalaSettings}
-  import uk.gov.hmrc.SbtAutoBuildPlugin
-  import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin
-  import uk.gov.hmrc.versioning.SbtGitVersioning
-  import play.sbt.routes.RoutesKeys.routesGenerator
-
-  val appName: String
-  val appDependencies : Seq[ModuleID]
-  lazy val plugins : Seq[Plugins] = Seq(play.sbt.PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin, SbtArtifactory)
-  lazy val playSettings : Seq[Setting[_]] = Seq.empty
-  lazy val scoverageSettings = {
-    import scoverage.ScoverageKeys
-    Seq(
-      ScoverageKeys.coverageExcludedPackages :=
-        "<empty>;Reverse.*;views.html.*;app.Routes.*;prod.*;uk.gov.hmrc.*;testOnlyDoNotUseInAppConf.*;forms.*;config.*;models.*;views.*",
-      ScoverageKeys.coverageMinimum := 99,
-      ScoverageKeys.coverageFailOnMinimum := true,
-      ScoverageKeys.coverageHighlighting := true
-    )
-  }
-  lazy val microservice: Project = Project(appName, file("."))
-    .enablePlugins(plugins : _*)
-    .settings(
-      playSettings,
-      routesGenerator := InjectedRoutesGenerator,
-      scalaSettings,
-      publishingSettings,
-      defaultSettings(),
-      scoverageSettings,
-      scalaVersion := "2.11.12",
-      libraryDependencies ++= appDependencies,
-      retrieveManaged := true,
-      evictionWarningOptions in update := EvictionWarningOptions.default.withWarnScalaVersionEviction(false),
-      majorVersion := 4
-    )
-    .settings(
-      resolvers += Resolver.bintrayRepo("hmrc", "releases"),
-      resolvers += Resolver.jcenterRepo
-    )
-}
